@@ -7,9 +7,9 @@ use LaravelArdent\Ardent\Ardent;
 class Log extends Ardent
 {
 
-    public static $rules = array(
+    public static $rules = [
         'locatie_id' => 'required|integer',
-    );
+    ];
 
 
     public function project()
@@ -48,5 +48,4 @@ class Log extends Ardent
     {
         return $this->hasMany('App\Models\Report')->where('completed', 0);
     }
-
 }
