@@ -11,7 +11,6 @@ use League\Flysystem\Filesystem;
 use League\Flysystem\Adapter\Local;
 use Nathanmac\Utilities\Parser\Facades\Parser;
 
-
 class HandleFloorplans extends Command
 {
     /**
@@ -48,8 +47,7 @@ class HandleFloorplans extends Command
         //lijst met te bewerken bestanden ophalen
         $floorplan = Floorplan::where('ready', 0)->first();
 
-        if(!empty($floorplan)){
-
+        if (!empty($floorplan)) {
             $project = Project::findOrFail($floorplan->project_id);
             $year = date("Y", strtotime($project->created_at));
 
@@ -60,34 +58,34 @@ class HandleFloorplans extends Command
             $file_location = $file_dir.$floorplan->filename;
 
             //eerst opruimen
-            if($filesystem->has($year.'/'.$floorplan->project_id.'/plattegrond/'.$floorplan->location_id.'/'.$floorplan->floor_id.'/plattegrond.png')){
+            if ($filesystem->has($year.'/'.$floorplan->project_id.'/plattegrond/'.$floorplan->location_id.'/'.$floorplan->floor_id.'/plattegrond.png')) {
                 $this->info('Plattegrond.png bestaat al, verwijderen');
                 $filesystem->delete($year.'/'.$floorplan->project_id.'/plattegrond/'.$floorplan->location_id.'/'.$floorplan->floor_id.'/plattegrond.png');
             }
             $filetype = $filesystem->getMimetype($year.'/'.$floorplan->project_id.'/plattegrond/'.$floorplan->location_id.'/'.$floorplan->floor_id.'/'.$floorplan->filename);
 
             //eerst checken of het een pdf is
-            if($filetype == 'application/pdf' || $filetype == 'image/jpeg') {
+            if ($filetype == 'application/pdf' || $filetype == 'image/jpeg') {
                 //pdf omzetten naar png
                 $this->info('PDF/JPG converteren naar PNG');
                 $this->info('convert -density 150 "'.$file_location.'" -quality 90 '.$file_dir.'plattegrond.png');
                 exec('convert -density 150 "'.$file_location.'" -quality 90 png8:'.$file_dir.'plattegrond.png');
             }
 
-            if($filesystem->has($year.'/'.$floorplan->project_id.'/plattegrond/'.$floorplan->location_id.'/'.$floorplan->floor_id.'/plattegrond.png')){
+            if ($filesystem->has($year.'/'.$floorplan->project_id.'/plattegrond/'.$floorplan->location_id.'/'.$floorplan->floor_id.'/plattegrond.png')) {
                 $this->info('Plattegrond.png gemaakt');
 
                 $this->info('gdal_translate -of vrt -expand rgba '.$file_dir.'plattegrond.png '.$file_dir.'temp.vrt');
 
                 exec('gdal_translate -of vrt -expand rgba '.$file_dir.'plattegrond.png '.$file_dir.'temp.vrt');
-            }else{
+            } else {
                 $this->info('Plattegrond.png niet gevonden.');
                 //@Todo: afhandeling
             }
 
             //eventuele folders met oud kaartmateriaal verwijderen
-            for($i = 0; $i < 7; $i++) {
-                if($filesystem->has($year.'/'.$floorplan->project_id.'/plattegrond/'.$floorplan->location_id.'/'.$floorplan->floor_id.'/'.$i)){
+            for ($i = 0; $i < 7; $i++) {
+                if ($filesystem->has($year.'/'.$floorplan->project_id.'/plattegrond/'.$floorplan->location_id.'/'.$floorplan->floor_id.'/'.$i)) {
                     $filesystem->deleteDir($year.'/'.$floorplan->project_id.'/plattegrond/'.$floorplan->location_id.'/'.$floorplan->floor_id.'/'.$i);
                     $this->info('Map '.$i.' verwijderd');
                 }
@@ -99,7 +97,7 @@ class HandleFloorplans extends Command
             //$this->info('Command: gdal2tiles.py -p raster -z 0-6 '.$file_dir.'plattegrond.png '.$file_dir);
             exec(base_path().'/gdal2tiles.py -l -p raster -z 0-6 '.$file_dir.'temp.vrt '.$file_dir);
 
-            if($filesystem->has($year.'/'.$floorplan->project_id.'/plattegrond/'.$floorplan->location_id.'/'.$floorplan->floor_id.'/0/0/0.png')){
+            if ($filesystem->has($year.'/'.$floorplan->project_id.'/plattegrond/'.$floorplan->location_id.'/'.$floorplan->floor_id.'/0/0/0.png')) {
                 //als dit bestand bestaat dan is de generatie succesvol geweest
                 $xml = $filesystem->read($year.'/'.$floorplan->project_id.'/plattegrond/'.$floorplan->location_id.'/'.$floorplan->floor_id.'/tilemapresource.xml');
                 $xml_array = Parser::xml($xml);
@@ -114,10 +112,6 @@ class HandleFloorplans extends Command
                 $floorplan->save();
                 $this->info('Done');
             }
-
-
-
         }
-
     }
 }
