@@ -29,11 +29,11 @@ class UserController extends Controller
         $users = User::all();
 
         return Datatables::of($users)
-            ->addColumn('action', function($object){
+            ->addColumn('action', function ($object) {
                 return '<a href="'.\URL::route('user.edit', ['id' => $object->id]).'"><i class="fa fa-search"></i></a>';
             })
-            ->addColumn('organisatie', function($object){
-                if(isset($object->client->naam)){
+            ->addColumn('organisatie', function ($object) {
+                if (isset($object->client->naam)) {
                     return $object->client->naam;
                 } else {
                     return 'Geen';
@@ -73,8 +73,7 @@ class UserController extends Controller
         $user->client_id = $request->input('client_id');
 
 
-        if(!$user->save()){
-
+        if (!$user->save()) {
             return redirect()->route('user.create')->with('errors', $user->errors());
         }
 
@@ -126,24 +125,23 @@ class UserController extends Controller
     {
         $user = User::findOrFail($id);
 
-        $user->first_name = \Input::get('first_name');
-        $user->last_name = \Input::get('last_name');
-        $user->email = \Input::get('email');
-        if(!empty(\Input::get('password'))) {
-            $user->password = \Input::get('password');
-            $user->password_confirmation = \Input::get('password_confirmation');
-        }else{
+        $user->first_name = $request->input('first_name');
+        $user->last_name = $request->input('last_name');
+        $user->email = $request->input('email');
+        if (!empty($request->input('password'))) {
+            $user->password = $request->input('password');
+            $user->password_confirmation = $request->input('password_confirmation');
+        } else {
             unset($user->password);
         }
-        $user->client_id = \Input::get('client_id');
+        $user->client_id = $request->input('client_id');
 
-        if(!$user->save()){
-
+        if (!$user->save()) {
             return redirect()->route('user.edit', ['id' => $user->id])->with('errors', $user->errors());
         }
 
-        if(!empty($request->input('role'))) {
-            $role = Role::find(\Input::get('role'));
+        if (!empty($request->input('role'))) {
+            $role = Role::find($request->input('role'));
             $user->detachRoles($user->roles);
             $user->attachRole($role);
 
@@ -151,7 +149,6 @@ class UserController extends Controller
         }
 
         return redirect()->route('home');
-
     }
 
     public function resetPassword()
@@ -159,26 +156,23 @@ class UserController extends Controller
         return \View::make('auth.forgot_password');
     }
 
-    public function storeResetPassword()
+    public function storeResetPassword(Request $request)
     {
-        $user = User::where('email', \Input::get('email'))->first();
+        $user = User::where('email', $request->input('email'))->first();
 
-        if($user != null) {
+        if ($user != null) {
             $new_password = str_random(10);
 
             $user->password = $new_password;
             $user->password_confirmation = $new_password;
 
             if ($user->save()) {
-
                 \Mail::send('emails.reminder', ['user' => $user, 'password' => $new_password], function ($m) use ($user) {
                     $m->from('info@davelaar.nl', 'Logboek Davelaarbouw B.V.');
 
                     $m->to($user->email, $user->name)->subject('Nieuw wachtwoord');
                 });
-
             }
-
         }
 
         return redirect()->route('login')->with('status', 'Nieuw wachtwoord verzonden.');
