@@ -3,15 +3,34 @@
 namespace App\Http\Middleware;
 
 use Closure;
-use Illuminate\Contracts\Auth\Guard;
+use Illuminate\Support\Facades\Auth;
 
 class Authenticate
 {
     /**
+     * Handle an incoming request.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @param  \Closure  $next
+     * @param  string|null  $guard
+     * @return mixed
+     */
+    public function handle($request, Closure $next, $guard = null)
+    {
+        if (Auth::guard($guard)->guest()) {
+            if ($request->ajax() || $request->wantsJson()) {
+                return response('Unauthorized.', 401);
+            } else {
+                return redirect()->guest('login');
+            }
+        }
+        return $next($request);
+    }
+    /**
      * The Guard implementation.
      *
      * @var Guard
-     */
+
     protected $auth;
 
     /**
@@ -19,7 +38,7 @@ class Authenticate
      *
      * @param  Guard  $auth
      * @return void
-     */
+
     public function __construct(Guard $auth)
     {
         $this->auth = $auth;
@@ -31,7 +50,7 @@ class Authenticate
      * @param  \Illuminate\Http\Request  $request
      * @param  \Closure  $next
      * @return mixed
-     */
+
     public function handle($request, Closure $next)
     {
         if ($this->auth->guest()) {
@@ -43,5 +62,5 @@ class Authenticate
         }
 
         return $next($request);
-    }
+    } */
 }
