@@ -19,14 +19,16 @@ class MenuMiddleware
 
         $user = \Auth::user();
 
-        if(!empty($user)) {
-            Menu::make('example', function($menu) use ($user) {
+        if (!empty($user)) {
+            Menu::make('example', function ($menu) use ($user) {
                 $menu->add('Home', '')->icon('fa fa-th')->active('');
                 $menu->add('Logboek', 'logboek')->icon('fa fa-archive')->active('projecten/*');
-                if($user->hasRole('admin')) $menu->logboek->add('Projecten', \URL::route('projecten.index'))->active('projecten');
+                if ($user->hasRole('admin')) {
+                    $menu->logboek->add('Projecten', \URL::route('projecten.index'))->active('projecten');
+                }
                 $menu->logboek->add('Rapporten', \URL::route('projecten.rapporten'))->active('projecten/rapporten');
 
-                if($user->hasRole(['admin', 'medewerker'])) {
+                if ($user->hasRole(['admin', 'medewerker'])) {
                     $menu->add('Subdatabase', 'subdatabase')->icon('fa fa-cubes')->active('subdatabase/*');
                     $menu->subdatabase->add('Klanten', 'subdatabase/client')->active('subdatabase/client');
                     $menu->subdatabase->add('Brandkleppen', 'subdatabase/firedamper')->active('subdatabase/firedamper');
@@ -37,7 +39,7 @@ class MenuMiddleware
                     $menu->subdatabase->add('QR-codes', 'subdatabase/generate-qr')->active('subdatbase/generate-qr');
                 }
 
-                if($user->hasRole('admin')) {
+                if ($user->hasRole('admin')) {
                     $menu->add('Gebruikers', 'gebruikers')->icon('fa fa-user')->active('user/*');
                     $menu->gebruikers->add('Gebruikers', 'user/')->active('user/');
                 }
