@@ -10,7 +10,8 @@ class Client extends Model
     use SoftDeletes;
     protected $dates = ['deleted_at'];
 
-    public function user() {
+    public function user()
+    {
         return $this->hasMany('App\User');
     }
 }
