@@ -35,14 +35,14 @@ class File extends Model
     public function location($clean = false)
     {
 
-        if(\Cache::has('project_created_at_'.$this->project_id)){
+        if (\Cache::has('project_created_at_'.$this->project_id)) {
             $created_at = \Cache::get('project_created_at_'.$this->project_id);
-        }else{
+        } else {
             $created_at = $this->project->created_at;
             \Cache::put('project_created_at_'.$this->project_id, $this->project->created_at, 10);
         }
 
-        if($clean) {
+        if ($clean) {
             return public_path().'/documenten/'.date("Y", strtotime($created_at)).'/'.$this->project_id.'/';
         }
 
@@ -51,20 +51,19 @@ class File extends Model
 
     public function thumbnail()
     {
-        if(\Cache::has('project_created_at_'.$this->project_id)){
+        if (\Cache::has('project_created_at_'.$this->project_id)) {
             $created_at = \Cache::get('project_created_at_'.$this->project_id);
-        }else{
+        } else {
             $created_at = $this->project->created_at;
             \Cache::put('project_created_at_'.$this->project_id, $this->project->created_at, 10);
         }
 
         $location = public_path().'/documenten/'.date("Y", strtotime($created_at)).'/'.$this->project_id.'/'.$this->id.'_thumb.jpg';
 
-        if(file_exists($location)) {
+        if (file_exists($location)) {
             return $location;
         }
 
         return public_path().'/documenten/'.date("Y", strtotime($created_at)).'/'.$this->project_id.'/'.$this->naam;
-
     }
 }
