@@ -1,8 +1,8 @@
 <?php
 
-return array(
+return [
 
-    'cache'      => array(
+    'cache'      => [
 
         /*
         |--------------------------------------------------------------------------
@@ -29,24 +29,24 @@ return array(
         | Cache settings
         |--------------------------------------------------------------------------
         */
-        'settings' => array(
+        'settings' => [
 
             'memoryCacheSize' => '32MB',
             'cacheTime'       => 600
 
-        ),
+        ],
 
         /*
         |--------------------------------------------------------------------------
         | Memcache settings
         |--------------------------------------------------------------------------
         */
-        'memcache' => array(
+        'memcache' => [
 
             'host' => 'localhost',
             'port' => 11211,
 
-        ),
+        ],
 
         /*
         |--------------------------------------------------------------------------
@@ -55,9 +55,9 @@ return array(
         */
 
         'dir'      => storage_path('cache')
-    ),
+    ],
 
-    'properties' => array(
+    'properties' => [
         'creator'        => 'Maatwebsite',
         'lastModifiedBy' => 'Maatwebsite',
         'title'          => 'Spreadsheet',
@@ -67,35 +67,35 @@ return array(
         'category'       => 'Excel',
         'manager'        => 'Maatwebsite',
         'company'        => 'Maatwebsite',
-    ),
+    ],
 
     /*
     |--------------------------------------------------------------------------
     | Sheets settings
     |--------------------------------------------------------------------------
     */
-    'sheets'     => array(
+    'sheets'     => [
 
         /*
         |--------------------------------------------------------------------------
         | Default page setup
         |--------------------------------------------------------------------------
         */
-        'pageSetup' => array(
+        'pageSetup' => [
             'orientation'           => 'portrait',
             'paperSize'             => '9',
             'scale'                 => '100',
             'fitToPage'             => false,
             'fitToHeight'           => true,
             'fitToWidth'            => true,
-            'columnsToRepeatAtLeft' => array('', ''),
-            'rowsToRepeatAtTop'     => array(0, 0),
+            'columnsToRepeatAtLeft' => ['', ''],
+            'rowsToRepeatAtTop'     => [0, 0],
             'horizontalCentered'    => false,
             'verticalCentered'      => false,
             'printArea'             => null,
             'firstPageNumber'       => null,
-        ),
-    ),
+        ],
+    ],
 
     /*
     |--------------------------------------------------------------------------
@@ -108,7 +108,7 @@ return array(
 
     'creator'    => 'Maatwebsite',
 
-    'csv'        => array(
+    'csv'        => [
         /*
        |--------------------------------------------------------------------------
        | Delimiter
@@ -135,9 +135,9 @@ return array(
         */
 
         'line_ending' => "\r\n"
-    ),
+    ],
 
-    'export'     => array(
+    'export'     => [
 
         /*
         |--------------------------------------------------------------------------
@@ -210,7 +210,7 @@ return array(
         | Default sheet settings
         |--------------------------------------------------------------------------
         */
-        'sheets'                      => array(
+        'sheets'                      => [
 
             /*
             |--------------------------------------------------------------------------
@@ -246,7 +246,7 @@ return array(
             |--------------------------------------------------------------------------
             */
             'strictNullComparison' => false
-        ),
+        ],
 
         /*
         |--------------------------------------------------------------------------
@@ -254,7 +254,7 @@ return array(
         |--------------------------------------------------------------------------
         */
 
-        'store'                       => array(
+        'store'                       => [
 
             /*
             |--------------------------------------------------------------------------
@@ -276,14 +276,14 @@ return array(
             */
             'returnInfo' => false
 
-        ),
+        ],
 
         /*
         |--------------------------------------------------------------------------
         | PDF Settings
         |--------------------------------------------------------------------------
         */
-        'pdf'                         => array(
+        'pdf'                         => [
 
             /*
             |--------------------------------------------------------------------------
@@ -298,48 +298,48 @@ return array(
             | PDF Driver settings
             |--------------------------------------------------------------------------
             */
-            'drivers' => array(
+            'drivers' => [
 
                 /*
                 |--------------------------------------------------------------------------
                 | DomPDF settings
                 |--------------------------------------------------------------------------
                 */
-                'DomPDF' => array(
+                'DomPDF' => [
                     'path' => base_path('vendor/dompdf/dompdf/')
-                ),
+                ],
 
                 /*
                 |--------------------------------------------------------------------------
                 | tcPDF settings
                 |--------------------------------------------------------------------------
                 */
-                'tcPDF'  => array(
+                'tcPDF'  => [
                     'path' => base_path('vendor/tecnick.com/tcpdf/')
-                ),
+                ],
 
                 /*
                 |--------------------------------------------------------------------------
                 | mPDF settings
                 |--------------------------------------------------------------------------
                 */
-                'mPDF'   => array(
+                'mPDF'   => [
                     'path' => base_path('vendor/mpdf/mpdf/')
-                ),
-            )
-        )
-    ),
+                ],
+            ]
+        ]
+    ],
 
-    'filters'    => array(
+    'filters'    => [
         /*
         |--------------------------------------------------------------------------
         | Register read filters
         |--------------------------------------------------------------------------
         */
 
-        'registered' => array(
+        'registered' => [
             'chunk' => 'Maatwebsite\Excel\Filters\ChunkReadFilter'
-        ),
+        ],
 
         /*
         |--------------------------------------------------------------------------
@@ -347,10 +347,10 @@ return array(
         |--------------------------------------------------------------------------
         */
 
-        'enabled'    => array()
-    ),
+        'enabled'    => []
+    ],
 
-    'import'     => array(
+    'import'     => [
 
         /*
         |--------------------------------------------------------------------------
@@ -415,12 +415,12 @@ return array(
         |--------------------------------------------------------------------------
         */
 
-        'encoding'                => array(
+        'encoding'                => [
 
             'input'  => 'UTF-8',
             'output' => 'UTF-8'
 
-        ),
+        ],
 
         /*
         |--------------------------------------------------------------------------
@@ -466,7 +466,7 @@ return array(
         |
         */
 
-        'dates'                   => array(
+        'dates'                   => [
 
             /*
             |--------------------------------------------------------------------------
@@ -490,15 +490,15 @@ return array(
             | Date columns
             |--------------------------------------------------------------------------
             */
-            'columns' => array()
-        ),
+            'columns' => []
+        ],
 
         /*
         |--------------------------------------------------------------------------
         | Import sheets by config
         |--------------------------------------------------------------------------
         */
-        'sheets'                  => array(
+        'sheets'                  => [
 
             /*
             |--------------------------------------------------------------------------
@@ -509,16 +509,16 @@ return array(
             |
             */
 
-            'test' => array(
+            'test' => [
 
                 'firstname' => 'A2'
 
-            )
+            ]
 
-        )
-    ),
+        ]
+    ],
 
-    'views'      => array(
+    'views'      => [
 
         /*
         |--------------------------------------------------------------------------
@@ -529,155 +529,155 @@ return array(
         |
         */
 
-        'styles' => array(
+        'styles' => [
 
             /*
             |--------------------------------------------------------------------------
             | Table headings
             |--------------------------------------------------------------------------
             */
-            'th'     => array(
-                'font' => array(
+            'th'     => [
+                'font' => [
                     'bold' => true,
                     'size' => 12,
-                )
-            ),
+                ]
+            ],
 
             /*
             |--------------------------------------------------------------------------
             | Strong tags
             |--------------------------------------------------------------------------
             */
-            'strong' => array(
-                'font' => array(
+            'strong' => [
+                'font' => [
                     'bold' => true,
                     'size' => 12,
-                )
-            ),
+                ]
+            ],
 
             /*
             |--------------------------------------------------------------------------
             | Bold tags
             |--------------------------------------------------------------------------
             */
-            'b'      => array(
-                'font' => array(
+            'b'      => [
+                'font' => [
                     'bold' => true,
                     'size' => 12,
-                )
-            ),
+                ]
+            ],
 
             /*
             |--------------------------------------------------------------------------
             | Italic tags
             |--------------------------------------------------------------------------
             */
-            'i'      => array(
-                'font' => array(
+            'i'      => [
+                'font' => [
                     'italic' => true,
                     'size'   => 12,
-                )
-            ),
+                ]
+            ],
 
             /*
             |--------------------------------------------------------------------------
             | Heading 1
             |--------------------------------------------------------------------------
             */
-            'h1'     => array(
-                'font' => array(
+            'h1'     => [
+                'font' => [
                     'bold' => true,
                     'size' => 24,
-                )
-            ),
+                ]
+            ],
 
             /*
             |--------------------------------------------------------------------------
             | Heading 2
             |--------------------------------------------------------------------------
             */
-            'h2'     => array(
-                'font' => array(
+            'h2'     => [
+                'font' => [
                     'bold' => true,
                     'size' => 18,
-                )
-            ),
+                ]
+            ],
 
             /*
             |--------------------------------------------------------------------------
             | Heading 2
             |--------------------------------------------------------------------------
             */
-            'h3'     => array(
-                'font' => array(
+            'h3'     => [
+                'font' => [
                     'bold' => true,
                     'size' => 13.5,
-                )
-            ),
+                ]
+            ],
 
             /*
              |--------------------------------------------------------------------------
              | Heading 4
              |--------------------------------------------------------------------------
              */
-            'h4'     => array(
-                'font' => array(
+            'h4'     => [
+                'font' => [
                     'bold' => true,
                     'size' => 12,
-                )
-            ),
+                ]
+            ],
 
             /*
              |--------------------------------------------------------------------------
              | Heading 5
              |--------------------------------------------------------------------------
              */
-            'h5'     => array(
-                'font' => array(
+            'h5'     => [
+                'font' => [
                     'bold' => true,
                     'size' => 10,
-                )
-            ),
+                ]
+            ],
 
             /*
              |--------------------------------------------------------------------------
              | Heading 6
              |--------------------------------------------------------------------------
              */
-            'h6'     => array(
-                'font' => array(
+            'h6'     => [
+                'font' => [
                     'bold' => true,
                     'size' => 7.5,
-                )
-            ),
+                ]
+            ],
 
             /*
              |--------------------------------------------------------------------------
              | Hyperlinks
              |--------------------------------------------------------------------------
              */
-            'a'      => array(
-                'font' => array(
+            'a'      => [
+                'font' => [
                     'underline' => true,
-                    'color'     => array('argb' => 'FF0000FF'),
-                )
-            ),
+                    'color'     => ['argb' => 'FF0000FF'],
+                ]
+            ],
 
             /*
              |--------------------------------------------------------------------------
              | Horizontal rules
              |--------------------------------------------------------------------------
              */
-            'hr'     => array(
-                'borders' => array(
-                    'bottom' => array(
+            'hr'     => [
+                'borders' => [
+                    'bottom' => [
                         'style' => 'thin',
-                        'color' => array('FF000000')
-                    ),
-                )
-            )
-        )
+                        'color' => ['FF000000']
+                    ],
+                ]
+            ]
+        ]
 
-    )
+    ]
 
-);
+];
