@@ -1,22 +1,22 @@
 <?php
 
-return array(
+return [
 
 
-    'pdf' => array(
+    'pdf' => [
         'enabled' => true,
         'binary'  => base_path('vendor/h4cc/wkhtmltopdf-amd64/bin/wkhtmltopdf-amd64'),
         'timeout' => false,
-        'options' => array(),
-        'env'     => array(),
-    ),
-    'image' => array(
+        'options' => [],
+        'env'     => [],
+    ],
+    'image' => [
         'enabled' => true,
         'binary'  => '/usr/local/bin/wkhtmltoimage',
         'timeout' => false,
-        'options' => array(),
-        'env'     => array(),
-    ),
+        'options' => [],
+        'env'     => [],
+    ],
 
 
-);
+];
