@@ -12,7 +12,7 @@ class AddEisAndWandVloerToLogs extends Migration
      */
     public function up()
     {
-        Schema::table('logs', function(Blueprint $t){
+        Schema::table('logs', function (Blueprint $t) {
             $t->integer('eis')->after('brandklep_id');
             $t->integer('oppervlak_type_id')->after('brandklep_id');
         });
@@ -25,7 +25,7 @@ class AddEisAndWandVloerToLogs extends Migration
      */
     public function down()
     {
-        Schema::table('logs', function(Blueprint $t){
+        Schema::table('logs', function (Blueprint $t) {
             $t->dropColumn('eis');
             $t->dropColumn('oppervlak_type_id');
         });
